+++ be/src/main/java/com/foodymoody/be/feed/domain/repository/FeedRepository.java
@@ -31,6 +31,4 @@ public interface FeedRepository {
 
     Optional<List<MenuNameRatingPair>> fetchMenuNameRatingList(List<ImageMenu> imageMenus);
 
-    Optional<Boolean> fetchIsLiked(FeedId feedId, MemberId memberId);
-
 }
