+++ be/src/main/java/com/foodymoody/be/feed/infra/persistence/jpa/FeedJpaRepository.java
@@ -32,11 +32,4 @@ public interface FeedJpaRepository extends JpaRepository<Feed, FeedId> {
             "WHERE im IN :imageMenus")
     Optional<List<MenuNameRatingPair>> fetchMenuNameRatingList(@Param("imageMenus") List<ImageMenu> imageMenus);
 
-    // TODO: 아마 여기서 isLiked가 계속 false로 들어오는 거임
-    @Query("SELECT fh.isLiked "
-            + "FROM FeedLike fh "
-            + "WHERE fh.feedId = :feedId "
-            + "AND fh.memberId = :memberId")
-    Optional<Boolean> fetchIsLiked(@Param("feedId") FeedId feedId, @Param("memberId") MemberId memberId);
-
 }
