+++ be/src/main/java/com/foodymoody/be/feed_like/infra/persistence/FeedLikeRepositoryImpl.java
@@ -1,5 +1,6 @@
 package com.foodymoody.be.feed_like.infra.persistence;
 
+import com.foodymoody.be.common.exception.NotFoundFeedHeartException;
 import com.foodymoody.be.common.util.ids.FeedId;
 import com.foodymoody.be.common.util.ids.MemberId;
 import com.foodymoody.be.feed_like.infra.persistence.jpa.FeedLikeJpaRepository;
@@ -27,6 +28,12 @@ public class FeedLikeRepositoryImpl implements FeedLikeRepository {
     @Override
     public void deleteByFeedIdAndMemberId(FeedId feedId, MemberId memberId) {
         feedLikeJpaRepository.deleteByFeedIdAndMemberId(feedId, memberId);
+    }
+
+    @Override
+    public boolean fetchIsLiked(FeedId feedId, MemberId memberId) {
+        return feedLikeJpaRepository.fetchIsLiked(feedId, memberId)
+                .orElse(false);
     }
 
 }
